fixed Faculty component, implemented through mapping, added faculty.ts
