+++ src/data/faculty.ts
@@ -1,0 +1,33 @@
+interface FacultyProps {
+  facultyName: string;
+  facultyImageLink: string;
+  facultyDescription: string;
+}
+
+export const faculty: FacultyProps[] = [
+  {
+    facultyName: "John Doe",
+    facultyImageLink: "https://i.nuuls.com/B66hZ.png",
+    facultyDescription: "Professor of Economics at UCR",
+  },
+  {
+    facultyName: "John Doe",
+    facultyImageLink: "https://i.nuuls.com/B66hZ.png",
+    facultyDescription: "Professor of Economics at UCR",
+  },
+  {
+    facultyName: "John Doe",
+    facultyImageLink: "https://i.nuuls.com/B66hZ.png",
+    facultyDescription: "Professor of Economics at UCR",
+  },
+  {
+    facultyName: "John Doe",
+    facultyImageLink: "https://i.nuuls.com/B66hZ.png",
+    facultyDescription: "Professor of Economics at UCR",
+  },
+  {
+    facultyName: "John Doe",
+    facultyImageLink: "https://i.nuuls.com/B66hZ.png",
+    facultyDescription: "Professor of Economics at UCR",
+  },
+];
